reviw: rename component and remove console.log
