+++ frontend/app/components/home/ExtremeCard.vue
